requirements.txt
aws-cdk-lib==2.96.2
constructs>=10.0.0,<11.0.0
boto3>=1.28.0
psycopg2-binary>=2.9.6